+++ backend/src/openarchiefbeheer/destruction/api/filtersets.py
@@ -96,6 +96,7 @@ class DestructionListItemFilterset(FilterSet):
 
 class DestructionListFilterset(FilterSet):
     name = CharFilter(lookup_expr="icontains")
+    comment = CharFilter(lookup_expr="icontains")
     status = ChoiceFilter(choices=ListStatus.choices)
     author = NumberFilter()
     reviewer = NumberFilter(
@@ -114,11 +115,23 @@ class DestructionListFilterset(FilterSet):
             assignees__user=value,
         )
 
-    ordering = OrderingFilter(fields=(("created", "created"), ("name", "name")))
+    ordering = OrderingFilter(
+        fields=(
+            ("name", "name"),
+            ("comment", "comment"),
+            ("created", "created"),
+            ("end", "end"),
+        )
+    )
 
     class Meta:
         model = DestructionList
-        fields = ("name", "status", "author", "reviewer", "assignee", "ordering")
+        fields = {
+            "comment": ["icontains"],
+            "created": ["exact", "gt", "lt", "gte", "lte", "isnull"],
+            "end": ["exact", "gt", "lt", "gte", "lte", "isnull"],
+            "status": ["exact"],
+        }
 
 
 class DestructionListReviewFilterset(FilterSet):