+++ frontend/src/pages/landing/Landing.loader.tsx
@@ -1,11 +1,9 @@
 import { User, whoAmI } from "../../lib/api/auth";
 import {
   DestructionList,
-  listDestructionLists,
+  getDestructionListsKanban,
 } from "../../lib/api/destructionLists";
 import { loginRequired } from "../../lib/auth/loaders";
-import { STATUS_MAPPING } from "../constants";
-import { STATUSES } from "./Landing";
 import "./Landing.css";
 
 export interface LandingContext {
@@ -15,11 +13,18 @@ export interface LandingContext {
 
 export const landingLoader = loginRequired(
   async ({ request }): Promise<LandingContext> => {
-    const abortController = new AbortController();
     const url = new URL(request.url);
+
     const urlSearchParams = url.searchParams;
-    const statusMap = await getStatusMap(urlSearchParams);
-    const user = await whoAmI(abortController.signal);
+    if (!urlSearchParams.has("ordering")) {
+      urlSearchParams.set("ordering", "-created");
+    }
+
+    const statusMap = await getDestructionListsKanban(
+      urlSearchParams,
+      request.signal,
+    );
+    const user = await whoAmI(request.signal);
 
     return {
       statusMap,
@@ -27,16 +32,3 @@ export const landingLoader = loginRequired(
     };
   },
 );
-
-export const getStatusMap = async (urlSearchParams: URLSearchParams) => {
-  if (!urlSearchParams.has("ordering")) {
-    urlSearchParams.set("ordering", "-created");
-  }
-  const lists = await listDestructionLists(urlSearchParams);
-  return STATUSES.reduce((acc, [status]) => {
-    const destructionLists = lists.filter(
-      (l) => STATUS_MAPPING[l.status] === status,
-    );
-    return { ...acc, [status || ""]: destructionLists };
-  }, {});
-};